+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Support for parsing playlist urls in web app (`https://www.youtube.com/playlist?list=ID`)
+
 ## [0.25.5] - 2024-09-06
 
 ### Fixed